Create Week 10 first test version
This section of code is now ready for the first round of testing. Although slightly jarring, the logic in the if-else clauses is functional and depicts the color of the sky quite well based on the data set. The next step is to clean up the code and integrate it into the main program.